requirement.txt
aiofiles==0.4.0
aiohttp==3.4.4
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.2.1
attrs==18.2.0
CacheControl==0.12.5
cached-property==1.4.2
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
firebase-admin==2.14.0
google-api-core==1.6.0
google-auth==1.6.1
google-cloud-core==0.28.1
google-cloud-firestore==0.30.0
google-cloud-storage==1.13.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.5
grpcio==1.17.0
httptools==0.0.11
idna==2.7
idna-ssl==1.1.0
Mako==1.0.7
MarkupSafe==1.1.0
mock==2.0.0
more-itertools==4.3.0
msgpack==0.6.0
multidict==4.5.2
pathlib2==2.3.3
pbr==5.1.1
pep8==1.7.1
pluggy==0.8.0
protobuf==3.6.1
py==1.7.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pytest==4.0.1
pytest-sanic==0.1.15
python-dateutil==2.7.5
python-dotenv==0.10.0
python-editor==1.0.3
python-firebase==1.2
pytz==2018.7
requests==2.20.1
rsa==4.0
sanic==0.8.3
six==1.11.0
ujson==1.35
urllib3==1.24.1
uvloop==0.11.3
websockets==5.0.1
yarl==1.2.0
